Renaming, reordering and refactoring some files
