+++ src/clients/artikkelit/actions/articleCheck.js
@@ -1,6 +1,7 @@
 import {idbGet} from '/artikkelit/utils/indexDB.js';
-import {enableElement, highlightElement, showSnackbar, startProcess, stopProcess} from '/common/ui-utils.js';
 import {validateArticleRecord} from '/common/rest.js';
+import {enableElement, highlightElement, showSnackbar, startProcess, stopProcess} from '/common/ui-utils.js';
+
