+++ src/clients/artikkelit/interfaces/publicationSearch.js
@@ -1,8 +1,9 @@
-import {getPublicationByISSN, getPublicationByISBN, getPublicationByTitle, getPublicationByMelindaId} from '/common/rest.js';
+import {resetAndHideCcLicense, showCcLicense} from '/artikkelit/interfaces/articleBasic.js';
 import {idbGet, idbClear, idbSet, idbGetStoredValues} from '/artikkelit/utils/indexDB.js';
-import {formToJson, setOptions, startProcess, stopProcess, showSnackbar} from '/common/ui-utils.js';
-import {showCcLicense, resetAndHideCcLicense} from '/artikkelit/interfaces/article.js';
 import {sortRecordData} from '/artikkelit/utils/utils.js';
+import {getPublicationByISSN, getPublicationByISBN, getPublicationByTitle, getPublicationByMelindaId} from '/common/rest.js';
+import {formToJson, setOptions, startProcess, stopProcess, showSnackbar} from '/common/ui-utils.js';
+
 
 export function initPublicationSearch(event) {
   console.log('initializing publication search...');
